requirements.txt
fastapi
uvicorn
pandas
google-generativeai
python-multipart